Make Example Makefile default rule compile all examples
The Example's default makefile rule now take a TARGET variable,
compiles it, runs on in the optimisations: O3, mergereturn (to
ensure there is a single ret command) and instnamer (to ensure
all instructions have names).

It then runs all the engines on the resulting bitcode.

It also generates a human-readable (-ish) file (.ll)