Add Debug flag and organise apron.cpp printing
Add a debug flag (-d) and organise apron.cpp to print additional
information when when the debug flag is set. The output is printed to
stderr. Note that the final output is not yet decided, and currenly many
modules within the project print out without regard to this flag (these
prints will be removed at a later commit)