Add AbstractState class
Add an Abstract State class which will hold all the abstract state
information in the pass. Currently, the planned information to be stored
is:
1. __user pointer may analysis (to which __user pointer a pointer 'may'
alias)
2. Apron integer analysis of integer variables (Currently left in
BasicBlock)
3. Apron integer analysis of read/write pointer range (i.e. to what
range within the pointer the code reads/writes).
4. Apron integer analysis of first0 range (i.e. in what range is the
first '\0' character in a pointer which is a string, e.g. for uses with
user_strncpy, user_strnlen, getcwd, or stat). Not yet implemented.