Update the AbstractState from a Call Instruction
Change CallValue to inherit from InstructionValue, and update the
BasicBlock's AbstarctState with the relevant information.

Currently handling the following kernel functions:
* get_user
* put_user
* copy_to_user
* copy_from_user

The following are tested but not implemented, or implementation is
incomplete:
* strnlen_user
* strncpy_from_user

Also added a method for debugging and testing:
* add_user_pointer

This method updates the AbstractState to mark that a pointer aliases a
__user pointer with the given offset. See kernel.c example file (next
commit) for more details and examples.